classificacao-service/requirements.txt
FastAPI==0.104.1
uvicorn==0.24.0
pydantic==2.10.0
pydantic-settings==2.3.0
SQLAlchemy==2.0.23
psycopg[binary]>=2.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
PyJWT==2.12.1
bcrypt==4.1.1
aio-pika==9.6.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
sqlalchemy-utils==0.41.1
python-dotenv==1.0.0
